Update sizeof code

+++ sizeof_int.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
 int main() {
-    printf("%lu\n", sizeof(int));
+    printf("%lu\n", sizeof(unsigned short int));
     return 0;
 }
